+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
+## 1.4.0
+1.修复popTo返回includeTargetFragment为true时，如果targetFragmentClass 为栈低时白屏问题
 
+2.修复popTo返回如果targetFragmentClass界面没有动画问题。
+
+3.修复动画问题。
+
+4.新增 Fragment setStackTag(String tag)， String getStackTag()，保存/获取堆栈信息
 ## 1.0.0-rc
 Support Androidx!